Ties php-http/guzzle6-adapter to pre-PHP raise version
Per [#182](https://github.com/SparkPost/php-sparkpost/issues/182), the php-http/guzzle6-adapter library, after not being updated for over a year decided to suddenly increase the PHP version requirements from 5.5 to 7.1 without issuing a new version (much less a new major release). This has caused existing composer checkouts to break if updating, and new ones to not work correctly on the affected PHP versions, which are supposed by this library.

This PR ties the composer requirements to the commit before the version increase, allowing for the fix of this issue.